refactor(api-page-builder): split prerendering plugins into impl and hooks
